requirements.txt
fastapi
jinja2
python-dotenv
asyncpg
python_socketio
jinja2